Refactor inventory module to separate API functionality
